refactor: useCardType -> useCardBrand로 이름 변경
