+++ hooks/src/lib/index.ts
@@ -2,7 +2,7 @@ export { default as useCardHolder } from './useCardHolder';
 export { default as useCVC } from './useCVC';
 export { default as usePassword } from './usePassword';
 export { default as useCardIssuer } from './useCardIssuer';
-export { default as useCardType } from './useCardType';
+export { default as useCardBrand } from './useCardBrand';
 export { default as useValidation } from './useValidation';
 export { default as useCardNumbers } from './useCardNumbers';
 export { default as useExpiryDate } from './useExpiryDate';
